server/requirements.txt
Django==4.0.4
django-registration==3.2
django-ulid==0.0.4
django-unrest==0.1.4
psycopg2==2.9.3
